Properly fill in nick names
Also edit the parser so timestamps become None when unspecified.
This allows the edit command to use the previous timestamp, and not
replace it with the current time, which causes unintentional reordering
of events.

Also, assign eventIDs incrementally if some already exist. This prevents
weird things like newly entered events getting old eventIDs.
But the timestamp is still supreme! Older events might have newer
eventIDs, but never will the case be that two events entered without
explicitly specifying eventIDs will have an unintuitive ordering with
respect to the eventIDs!